+++ src/main/java/org/ojalgo/array/Array1D.java
@@ -52,8 +52,7 @@ import org.ojalgo.tensor.TensorFactory1D;
  *
  * @author apete
  */
-public final class Array1D<N extends Comparable<N>> extends AbstractList<N>
-        implements Access1D.Visitable<N>, Access1D.Aggregatable<N>, Access1D.Sliceable<N>, Access1D.Elements, Access1D.IndexOf,
+public final class Array1D<N extends Comparable<N>> extends AbstractList<N> implements Access1D.Visitable<N>, Access1D.Aggregatable<N>, Access1D.Sliceable<N>,
         Access1D.Collectable<N, Mutate1D>, Mutate1D.ModifiableReceiver<N>, Mutate1D.Mixable<N>, Mutate1D.Sortable, RandomAccess {
 
     public static final class Factory<N extends Comparable<N>>
@@ -143,7 +142,7 @@ public final class Array1D<N extends Comparable<N>> extends AbstractList<N>
 
             long i = low, j = high;
 
-            final double pivot = myArray.doubleValue(low + (high - low) / 2);
+            double pivot = myArray.doubleValue(low + (high - low) / 2);
 
             while (i <= j) {
 
@@ -206,7 +205,7 @@ public final class Array1D<N extends Comparable<N>> extends AbstractList<N>
 
             long i = low, j = high;
 
-            final double pivot = myArray.doubleValue(low + (high - low) / 2);
+            double pivot = myArray.doubleValue(low + (high - low) / 2);
 
             while (i <= j) {
 
@@ -393,7 +392,7 @@ public final class Array1D<N extends Comparable<N>> extends AbstractList<N>
 
     @Override
     public int hashCode() {
-        final int prime = 31;
+        int prime = 31;
         int result = super.hashCode();
         result = prime * result + (int) (length ^ length >>> 32);
         result = prime * result + (myDelegate == null ? 0 : myDelegate.hashCode());
@@ -405,7 +404,7 @@ public final class Array1D<N extends Comparable<N>> extends AbstractList<N>
 
     @Override
     public int indexOf(final Object obj) {
-        final int tmpLength = this.size();
+        int tmpLength = this.size();
         if (obj == null) {
             for (int i = 0; i < tmpLength; i++) {
                 if (this.get(i) == null) {
@@ -424,25 +423,10 @@ public final class Array1D<N extends Comparable<N>> extends AbstractList<N>
 
     @Override
     public long indexOfLargest() {
-        return this.indexOfLargestInRange(myFirst, myLimit);
-    }
-
-    @Override
-    public long indexOfLargestInRange(final long first, final long limit) {
-        return (myDelegate.indexOfLargest(this.convert(first), this.convert(limit), myStep) - myFirst) / myStep;
-    }
-
-    /**
-     * @see Scalar#isAbsolute()
-     */
-    @Override
-    public boolean isAbsolute(final long index) {
-        return myDelegate.isAbsolute(this.convert(index));
-    }
-
-    @Override
-    public boolean isAllSmall(final double comparedTo) {
-        return myDelegate.isSmall(myFirst, myLimit, myStep, comparedTo);
+        long first = this.convert(myFirst);
+        long limit = this.convert(myLimit);
+        long step = myStep;
+        return (myDelegate.indexOfLargest(first, limit, step) - first) / step;
     }
 
     @Override
@@ -450,20 +434,12 @@ public final class Array1D<N extends Comparable<N>> extends AbstractList<N>
         return length == 0;
     }
 
-    /**
-     * @see Scalar#isSmall(double)
-     */
-    @Override
-    public boolean isSmall(final long index, final double comparedTo) {
-        return myDelegate.isSmall(this.convert(index), comparedTo);
-    }
-
     @Override
     public double mix(final long index, final BinaryFunction<N> mixer, final double addend) {
         ProgrammingError.throwIfNull(mixer);
         synchronized (myDelegate) {
-            final double oldValue = this.doubleValue(index);
-            final double newValue = mixer.invoke(oldValue, addend);
+            double oldValue = this.doubleValue(index);
+            double newValue = mixer.invoke(oldValue, addend);
             this.set(index, newValue);
             return newValue;
         }
@@ -473,8 +449,8 @@ public final class Array1D<N extends Comparable<N>> extends AbstractList<N>
     public N mix(final long index, final BinaryFunction<N> mixer, final N addend) {
         ProgrammingError.throwIfNull(mixer);
         synchronized (myDelegate) {
-            final N oldValue = this.get(index);
-            final N newValue = mixer.invoke(oldValue, addend);
+            N oldValue = this.get(index);
+            N newValue = mixer.invoke(oldValue, addend);
             this.set(index, newValue);
             return newValue;
         }
@@ -652,13 +628,13 @@ public final class Array1D<N extends Comparable<N>> extends AbstractList<N>
 
         if (myDelegate.isPrimitive()) {
 
-            final double tmpVal = this.doubleValue(indexA);
+            double tmpVal = this.doubleValue(indexA);
             this.set(indexA, this.doubleValue(indexB));
             this.set(indexB, tmpVal);
 
         } else {
 
-            final N tmpVal = this.get(indexA);
+            N tmpVal = this.get(indexA);
             this.set(indexA, this.get(indexB));
             this.set(indexB, tmpVal);
         }
@@ -672,7 +648,7 @@ public final class Array1D<N extends Comparable<N>> extends AbstractList<N>
 
         long i = low, j = high;
 
-        final double pivot = this.doubleValue(low + (high - low) / 2);
+        double pivot = this.doubleValue(low + (high - low) / 2);
 
         while (i <= j) {
 
@@ -702,7 +678,7 @@ public final class Array1D<N extends Comparable<N>> extends AbstractList<N>
 
         long i = low, j = high;
 
-        final double pivot = this.doubleValue(low + (high - low) / 2);
+        double pivot = this.doubleValue(low + (high - low) / 2);
 
         while (i <= j) {